- larger use of forware declarations : should compile faster - corrected optobjective view problem - corected find function of MOVector - Problemthreads are terminated when application quitted
git-svn-id: https://openmodelica.org/svn/OpenModelica/trunk@11484 f25d12d1-65f4-0310-ae8a-bbce733d8d8e

+++ OMOptim/OMOptimBasis/Problems/OMCase.cpp
@@ -55,8 +55,6 @@ OMCase::OMCase(ProjectBase* project)
 
 OMCase::OMCase(const OMCase &omCase)
 {
-
-
     _name = omCase._name;
     _project = omCase._project;
 
@@ -158,11 +156,11 @@ void OMCase::store(QString destFolder, QString tempDir)
     {
         dir.mkpath(_saveFolder);
     }
-    else
-    {
-        LowTools::removeDir(_saveFolder);
-        dir.mkdir(_saveFolder);
-    }
+//    else
+//    {
+//        LowTools::removeDir(_saveFolder);
+//        dir.mkdir(_saveFolder);
+//    }
 
     //copy needed path from old place to new one
     if(tempDir != "")
@@ -218,7 +216,8 @@ void OMCase::rename(QString newName, bool changeFolder)
         }
 
         LowTools::copyDir(oldSaveFolder,newSaveFolder);
-        LowTools::removeDir(oldSaveFolder);
+        if(oldSaveFolder!=newSaveFolder)
+            LowTools::removeDir(oldSaveFolder);
 
         newDir.rename(oldSaveFileName,newSaveFileName);
